requirements.txt
asgiref==3.11.0
black==25.11.0
click==8.3.1
colorama==0.4.6
Django==5.2.8
djangorestframework==3.16.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pytokens==0.3.0
sqlparse==0.5.3
tzdata==2025.2
django-autoslug==1.9.9